all i could do for now
from what i remember, then there was a fundamental issue with the tests. Some of the test cases in themselves does not make sense and others agreed on discord.
I remember it as this would break some old tests and fix the issue i had, but i did not like the idea of breaking old tests without full proof that breaking the old tests was a "good" thing.